+++ script/detach.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
-rm -rf .git CHANGELOG.md LICENSE README.md composer.json
+rm -rf .git CHANGELOG.md LICENSE README.md
 git init