+++ public/index.php
@@ -4,5 +4,4 @@
  * Start application
  */
 
-require_once("../vendor/autoload.php");
 require_once("../App/app.php");