+++ App/Layout/Head.php
@@ -33,15 +33,8 @@ final class Head extends \Stateless\Layout {
             "<!DOCTYPE html><html><head>" .
             "<title>" . $this->title . "</title>" .
             "<meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">" .
-            "<link href=\"https://bootswatch.com/4/pulse/bootstrap.min.css\" " .
-                "rel=\"stylesheet\" type=\"text/css\">" .
-            "<link href=\"/css/style.css\" rel=\"stylesheet\" type=\"text/css\">" .
-            "<script
-                src=\"https://code.jquery.com/jquery-3.3.1.js\"
-                integrity=\"sha256-2Kok7MbOyxpgUVvAk/HJ2jigOSYS2auK4Pfzbm7uH60=\"
-                crossorigin=\"anonymous\"></script>" .
-            "<script src=\"https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/js/bootstrap.min.js\">" .
-                "</script>" .
+            "<link href=\"/css/main.css\" rel=\"stylesheet\" type=\"text/css\">" .
+            "<script src=\"/js/main.js\"></script>" .
             "</head><body";
 
         if (!empty($this->class)) {