+++ script/build.sh
@@ -1,0 +1,7 @@
+# Build SASS
+echo "Building SASS...";
+npm run sass
+
+# Webpack js
+echo "Building JS...";
+npm run webpack