+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [Issue #5] - Update project title in composer.json
 - [Issue #3] - Detach script should not delete composer.json
+- [Issue #2] - App should not assume Composer packages are installed
 - [Issue #1] - Vendor autoloader appears in index.php and app.php
 
 ## [0.0.1] 2018-02-24